Fix INT1 rate (again)
By increasing the system clock step (number of CPU clocks we emulate
before checking all hardware), the INT1 rate is closer to what is
supposed to be. Without a proper scheduler implementation, this will
never be really accurate so this value is effectively a magic number
that might or might not work dependng on the software being emulated.